api/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# Tableau Authentication
PyJWT>=2.8.0
cryptography>=41.0.0

# ML/AI Features
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# Caching & Performance
cachetools>=5.3.0

# Utilities
python-multipart>=0.0.6
